[TRTLLM-15177][chore] Kimi K3: inline MLA module into modeling_kimi_linear
Move the K3-specific KimiK3MLAAttention (and its rope/dtype helpers) from
the modules/kimi_k3_mla package directly into modeling_kimi_linear.py,
matching the per-model modeling_xxx.py convention (e.g. DeepSeek-V3), and
delete the now-empty package. The moved code is unchanged apart from
following the file's Linear-as-TrtllmLinear alias. kimi_kda stays a
standalone module (general enough to warrant it).

Deferred cleanup from PR #17269, tracked in TRTLLM-15177. Supersedes
PR #18159, which inlined the pre-refactor version of the module and no
longer rebases cleanly after the shared-MLA output-gate-hook refactor
(#18375) and attention-domain consolidation (#17968).
